backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bidict==0.22.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
pydantic==2.5.3
pydantic_core==2.14.6
python-engineio==4.8.2
python-socketio==5.11.0
simple-websocket==1.0.0
sniffio==1.3.0
sockets==1.0.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0
wsproto==1.2.0
